Make record data non-nullable
